Big update and cleanup
Apart from lots of tweaks, the main change is `localize_config'. This
changes the configuration of Polybar to adapt to the available screens
and modules.

Next step would be to include scripts to update this git repo with local
updates and to update local files from the git repo in an easy way.