Use blurred screengrab for locking
This is so a lock screen image is displayed correctly on a multi-monitor
setup.
In the future I might add a way to create a full sized background image
from a wallpaper, so it can also be shown correctly on a multi-monitor
setup (with different resolution per monitor). That way I could also
maintain just 1 high res library of 'lock pictures' that get scaled
down/up to fit on the current monitor(s).